Separate the table from employee and client, create a 'super component' in layout
